+++ hutool-cache/src/test/java/cn/hutool/cache/test/CacheTest.java
@@ -58,9 +58,11 @@ public class CacheTest {
 		lruCache.get("key1");
 		lruCache.put("key4", "value4", DateUnit.SECOND.getMillis() * 3);
 		
+		String value1 = lruCache.get("key1");
+		Assert.assertNotNull(value1);
 		//由于缓存容量只有3，当加入第四个元素的时候，根据LRU规则，最少使用的将被移除（2被移除）
 		String value2 = lruCache.get("key2");
-		Assert.assertTrue(null == value2);
+		Assert.assertNull(value2);
 	}
 	
 	@Test