+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 * 【poi】         CellUtil.getCellValue增加null验证
 * 【core】       增加文件内容跟随器Tailer
 * 【crypto】    增加RC4算法
+* 【core】       增加FixedLinkedHashMap
+* 【extra】       增加ChannelType，JschUtil增加createSession、createChannel、openChannel等方法
 
 ### Bug修复
 * 【core】       修复ExceptionUtil（pr#35@Gitee）
@@ -21,6 +23,7 @@
 * 【core】       修复TimedCache中onRemove失效问题（issue#ITD0O@Gitee）
 * 【core】       修复DateConverter日期负数问题（issue#ITWK4@Gitee）
 * 【json】       修复toBean时父类定义泛型字段导致的注入问题（issue#ITGGN@Gitee）
+* 【cahce】    修复读锁导致的LRU异常（issue#303@Gtihub）
 
 -------------------------------------------------------------------------------------------------------------